+++ app/Source/RideRequest/Domain/CancelRide/CancelRideLogic.php
@@ -8,6 +8,7 @@ use App\Source\RideRequest\Infra\CancelRide\Services\CancelRideService;
 use App\Source\RideRequest\Infra\CancelRide\Services\RatingService;
 use App\Source\RideRequest\Infra\CancelRide\Specifications\CanCancelRideSpecification;
 use App\Source\RideRequest\Infra\CancelRide\Specifications\IsLateCancellationSpecification;
+use App\Source\RideRequest\Infra\Common\Services\UpdatePendingRequestsCountService;
 use Exception;
 
 class CancelRideLogic
@@ -16,7 +17,8 @@ class CancelRideLogic
         private readonly CanCancelRideSpecification $canCancelRideSpecification,
         private readonly CancelRideService $cancelRideService,
         private readonly RatingService $ratingService,
-        private readonly IsLateCancellationSpecification $isLateCancellationSpecification
+        private readonly IsLateCancellationSpecification $isLateCancellationSpecification,
+        private readonly UpdatePendingRequestsCountService $updatePendingRequestsCountService
     ) {
     }
 
@@ -37,18 +39,24 @@ class CancelRideLogic
             throw new Exception('You cannot cancel this ride');
         }
 
-        $rideRequest = $this->cancelRideService->cancel(
-            rideRequest: $rideRequest,
-            authUserId: $authUserId
-        );
-
-        if ($this->isLateCancellationSpecification->isSatisfied($ride)) {
+        if ($this->isLateCancellationSpecification->isSatisfied($rideRequest)) {
             $this->ratingService->setLateCancellationRating($rideRequest);
             $this->cancelRideService->increaseLastMinuteCancellation($authUserId);
         } else {
             $this->ratingService->remove($rideRequest);
-            $this->cancelRideService->decreaseRidesCount($rideRequest);
+            $this->cancelRideService->decreaseRidesCountForPassenger($rideRequest);
         }
+
+        //driver can only reject/accept and then cancel, when driver accepts/rejects count will be decreased
+        //but here if passenger cancels while request is pending then we need to decrease count
+        if ($rideRequest->amIPassenger() && $rideRequest->isPending()) {
+            $this->updatePendingRequestsCountService->decreaseForDriver($ride);
+        }
+
+        $rideRequest = $this->cancelRideService->cancel(
+            rideRequest: $rideRequest,
+            authUserId: $authUserId
+        );
 
         NotifyUserLogic::sendCancellationNotification($rideRequest, $authUserId);