+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.21.0] - 2023-04-22
+
+### Added
+
+- Passenger cannot request multiple rides for the same route and the same date
+- Passenger can now cancel pending ride
+    - If passenger does that decrease driver's pending rides count
+
+### Changed
+
+- Late cancellation can happen if ride request is accepted and if ride is active
+
 ## [1.20.2] - 2023-04-21
 
 ### Added